+++ .eslintrc.js
@@ -1,3 +1,7 @@
 module.exports = {
   extends: ["plugin:@vetan2/ts"],
+  plugins: ["@vetan2"],
+  rules: {
+    "@vetan2/import-path": 2,
+  },
 };