+++ packages/eslint-plugin/src/index.ts
@@ -1,7 +1,7 @@
 import typescriptConfig from "@vetan2/eslint-config-ts";
 import baseConfig from "@vetan2/eslint-config-base";
 import reactConfig from "@vetan2/eslint-config-react";
-import type { ESLint } from "eslint";
+import rules from "./rules";
 
 export = {
   configs: {
@@ -9,4 +9,5 @@ export = {
     ts: typescriptConfig,
     react: reactConfig,
   },
-} satisfies ESLint.Plugin;
+  rules,
+};