+++ packages/eslint-plugin/src/rules/index.ts
@@ -1,0 +1,5 @@
+import importPath from "./import-path";
+
+export default {
+  "import-path": importPath,
+};