+++ packages/eslint-plugin/src/utils.ts
@@ -1,0 +1,4 @@
+import { ESLintUtils } from "@typescript-eslint/utils";
+
+// TODO docs url
+export const createRule = ESLintUtils.RuleCreator(() => "");