+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Add `import-path` rule to plugin
+
 ## 1.4.0 - 2023-08-19
 
 ## 1.3.0 - 2023-08-13